requirements.txt
gspread
oauth2client
pandas
google-cloud-bigquery
dbt-bigquery